+++ src/lib/builders/index.ts
@@ -13,4 +13,5 @@ export * from './slider';
 export * from './dialog';
 export * from './tooltip';
 export * from './pagination';
+export * from './toolbar';
 export * from './dropdown-menu';