+++ src/routes/docs/builders/accordion/preview.svelte
@@ -24,8 +24,8 @@
 	];
 </script>
 
-<Docs.PreviewWrapper>
-	<div class="w-full rounded-md bg-[--line-color] shadow-lg" {...root}>
+<Docs.PreviewWrapper fullwidth>
+	<div class="mx-auto w-full max-w-md rounded-md bg-[--line-color] shadow-lg" {...root}>
 		{#each items as accItem}
 			<div {...$item(accItem.id)} class="accordion-item">
 				<h2 class="flex">
@@ -42,10 +42,6 @@
 </Docs.PreviewWrapper>
 
 <style lang="postcss">
-	.comp-preview {
-		--line-color: theme('colors.gray.300');
-	}
-
 	.accordion-item {
 		@apply mt-px overflow-hidden first:mt-0 first:rounded-t last:rounded-b 
 				focus-within:relative focus-within:z-10 focus-within:ring focus-within:ring-magnum-400;
@@ -54,7 +50,7 @@
 	.accordion-trigger {
 		@apply flex h-12 flex-1  cursor-pointer items-center
 				justify-between bg-white px-5 text-base font-medium leading-none text-magnum-700 shadow-[0_1px_0]
-				shadow-[--line-color] outline-none hover:bg-gray-200;
+				shadow-[--line-color]  hover:bg-gray-200;
 	}
 
 	.accordion-content {